# src/COP.API/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.1.0",
    "title" : "MONICA COP API"
  },
  "host" : "portal.monica-cloud.eu",
  "basePath" : "/cop/api",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/person" : {
      "post" : {
        "tags" : [ "Person" ],
        "description" : "add a new agent",
        "parameters" : [ {
          "in" : "body",
          "name" : "fullName",
          "description" : "Users full name",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Person"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/persons" : {
      "get" : {
        "tags" : [ "Person" ],
        "description" : "Get all agents",
        "parameters" : [ {
          "name" : "role",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "enum" : [ "POLICE", "GUARD", "CREW", "PARAMEDICS" ]
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GetPersonListResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/person/{id}" : {
      "get" : {
        "tags" : [ "Person" ],
        "description" : "get agents by id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Person" ],
        "description" : "update person",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "updates",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Person"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Person" ],
        "description" : "delete person by id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Person" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "description" : "Person id"
        },
        "fullName" : {
          "type" : "string",
          "description" : "Person name"
        },
        "phone" : {
          "type" : "string",
          "description" : "Person phone number"
        },
        "email" : {
          "type" : "string",
          "description" : "Person email address"
        },
        "role" : {
          "type" : "string",
          "description" : "Person role (e.g. GUARD, CREW, PARAMEDIC,POLICE)"
        },
        "active" : {
          "type" : "boolean",
          "description" : "Person active? "
        }
      },
      "example" : {
        "role" : "role",
        "phone" : "phone",
        "fullName" : "fullName",
        "active" : true,
        "id" : 0,
        "email" : "email"
      }
    },
...
    "GetPersonListResponse" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/Person"
      }
    },
...
    "GeneralResponse" : {
      "type" : "object",
      "required" : [ "description", "success" ],
      "properties" : {
        "success" : {
          "type" : "boolean",
          "description" : "returns 1 if successful"
        },
        "description" : {
          "type" : "string",
          "description" : "a short comment"
        }
      },
      "example" : {
        "success" : true,
        "description" : "description"
      }
    },
    "ErrorResponse" : {
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}